+++ xray-svn-trunk/xr_3da/xrGame/WeaponShotgun.cpp
@@ -353,9 +353,9 @@ u8 CWeaponShotgun::AddCartridge		(u8 cnt)
 {
 	if(IsMisfire())	bMisfire = false;
 
-	if(m_set_next_ammoType_on_reload != u32(-1)){
+	if(m_set_next_ammoType_on_reload != u8(-1)){
 		m_ammoType						= m_set_next_ammoType_on_reload;
-		m_set_next_ammoType_on_reload	= u32(-1);
+		m_set_next_ammoType_on_reload	= u8(-1);
 
 	}
